Optimizing equivalence queries
I realized we can lean on BigQuery more for the equivalence checks.
Instead of doing one query on the per-month tables and one equivalent
query on the per-project table, then comparing results, we can do it all
in a single BigQuery query.

This changes the queries so that they only return rows if there is a
mismatch in test_ids between the two tables.

This allows faster processing because it's immediately half the total
queries and we can do queries over larger time windows because BigQuery
should be able to handle these queries more efficiently (less data to
send down to the client).