Relax requirements around date range checking
We currently check whether users specify a start time window that
precedes M-Lab epoch, which is 2009-02-11 because we don't have
per-month tables that precede 2009-02, but we shouldn't reject it if the
user specifies something like --start_date=2009-02-01 because we can
still query over the 2009_02 table.

This change relaxes the time window validation and accepts any start
time window going back to 2009-02-01.

Additionally this improves the error message when a user inputs an invalid
time value to print out what the expected range is.