Trims the size of check failure messages when there are many test_ids
The list of mismatched test_id values can get very long (like thousands
of entries long for a single check). It's not useful to print the full
list to the console, so this adds intelligence to trim down the list
into a manageable size of no more than 10 items, with duplicates
removed.